feat: Refactor database tables and types for improved data handling and consistency
- Updated License, Settings, Users, Shifts, Categories, Orders, and OrderItems tables to align with new entity structures.
- Enhanced DateTimeConverter to enforce UTC and local time handling.
- Refactored OrderCalculatorService to utilize a CalculableItem interface for better abstraction.
- Improved reports data models to include net sales and total collected amounts.
- Added JSON serialization methods to report models for cloud sync compatibility.
- Enhanced AppSettingsModel with safe JSON parsing and default values to prevent crashes.
- Updated shift calculations to accurately reflect sales, refunds, and expenses during shifts.

+++ lib/features/reports/domain/entities/item_sales_entity.dart
@@ -3,7 +3,7 @@ import 'package:equatable/equatable.dart';
 class ItemSalesEntity extends Equatable {
   final String itemName;
   final int quantitySold;
-  final double totalRevenue;
+  final int totalRevenue; 
 
   const ItemSalesEntity({
     required this.itemName,
